Fix multiple server crashes
