+++ src/main/java/com/rewindmc/retroemi/RetroEMI.java
@@ -18,9 +18,6 @@ import net.minecraft.client.renderer.RenderBlocks;
 import net.minecraft.client.renderer.entity.RenderItem;
 import net.minecraft.client.resources.I18n;
 import net.minecraft.item.ItemBlock;
-import net.minecraft.nbt.NBTTagCompound;
-import net.minecraft.nbt.NBTTagList;
-import org.jetbrains.annotations.Nullable;
 import org.lwjgl.opengl.GL11;
 import shim.com.mojang.blaze3d.systems.RenderSystem;
 import it.unimi.dsi.fastutil.ints.IntOpenHashSet;
@@ -40,7 +37,6 @@ import net.minecraft.client.gui.GuiScreen;
 import net.minecraft.client.gui.ScaledResolution;
 import net.minecraft.client.gui.inventory.GuiContainer;
 import net.minecraft.client.renderer.Tessellator;
-import net.minecraft.entity.player.EntityPlayer;
 import net.minecraft.item.Item;
 import net.minecraft.item.ItemPotion;
 import net.minecraft.item.ItemStack;
@@ -52,7 +48,6 @@ import shim.net.minecraft.client.gui.tooltip.TooltipPositioner;
 import shim.net.minecraft.client.util.ITooltipFlag;
 import shim.net.minecraft.client.util.math.MatrixStack;
 import shim.net.minecraft.client.util.math.Vec2i;
-import shim.net.minecraft.item.ItemStacks;
 import shim.net.minecraft.text.MutableText;
 import shim.net.minecraft.text.Text;
 import shim.net.minecraft.util.Formatting;
@@ -60,8 +55,6 @@ import shim.net.minecraft.util.Formatting;
 public class RetroEMI {
 	public static final RetroEMI instance = new RetroEMI();
 	public final RenderItem itemRenderer;
-
-	private static final List<Runnable> tickQueue = new ArrayList<>();
 
 	private RetroEMI() {
 		if (!FMLCommonHandler.instance().getSide().isServer()) {
@@ -79,81 +72,11 @@ public class RetroEMI {
 		return false;
 	}
 
-	public static void executeOnMainThread(Runnable r) {
-		synchronized (tickQueue) {
-			tickQueue.add(r);
-		}
-	}
-
-	public static void tick() {
-		Runnable[] queue;
-		synchronized (tickQueue) {
-			queue = tickQueue.toArray(new Runnable[tickQueue.size()]);
-			tickQueue.clear();
-		}
-		for (Runnable r : queue) {
-			r.run();
-		}
-	}
-
 	public static Collection<PotionEffect> getEffects(EmiStack stack) {
 		if (stack.getItemStack().getItem() instanceof ItemPotion p) {
 			return p.getEffects(stack.getItemStack());
 		}
 		return Collections.emptyList();
-	}
-
-	public static List<String> wrapLines(String str, int cols) {
-		ArrayList<String> li = new ArrayList<String>();
-		StringBuilder buf = new StringBuilder();
-		for (String line : str.split("\n")) {
-			int w = -1;
-			for (String word : line.split(" ")) {
-				if (w + 1 + word.length() > cols) {
-					li.add(buf.toString());
-					buf.setLength(0);
-					w = 0;
-				} else {
-					if (w != -1) buf.append(" ");
-					w++;
-				}
-				while (word.length() > cols) {
-					li.add(word.substring(0, cols));
-					word = word.substring(cols);
-				}
-				buf.append(word);
-				w += word.length();
-			}
-			if (buf.length() > 0) {
-				li.add(buf.toString());
-			}
-			buf.setLength(0);
-		}
-		return li;
-	}
-
-	public static String join(List<String> strs, String delim) {
-		StringBuilder sb = new StringBuilder();
-		boolean first = true;
-		for (String s : strs) {
-			if (first) {
-				first = false;
-			} else {
-				sb.append(delim);
-			}
-			sb.append(s);
-		}
-		return sb.toString();
-	}
-
-	public static void offerOrDrop(EntityPlayer player, ItemStack stack) {
-		if (!player.inventory.addItemStackToInventory(stack)) {
-			player.dropPlayerItemWithRandomChoice(stack, false);
-		}
-	}
-
-	public static boolean canCombine(ItemStack a, ItemStack b) {
-		return a == null || b == null ? a == b : a.isItemEqual(b) && ItemStack.areItemStackTagsEqual(a, b);
 	}
 
 	public static void renderModernTooltip(GuiScreen screen, List<TooltipComponent> components, int x, int y, int maxWidth, TooltipPositioner positioner) {
@@ -290,10 +213,6 @@ public class RetroEMI {
 		return !I18n.format(s).equals(s);
 	}
 
-	public static String replaceCharAt(String s, int index, char c) {
-		return s.substring(0, index) + c + s.substring(index + 1);
-	}
-
 	public static List<Item> getAllItems() {
 		List<Item> items = new ArrayList<>();
 		((Iterable<Item>) EmiPort.getItemRegistry()).forEach(items::add);
@@ -307,36 +226,6 @@ public class RetroEMI {
 	public static int getScaledWidth(Minecraft client) {
 		return client.displayWidth / EmiPort.getGuiScale(client);
 	}
-
-	private static @Nullable String getIdInner(ItemStack stack) {
-		if (ItemStacks.isEmpty(stack)) {
-			return null;
-		}
-		Item item = stack.getItem();
-		if (item instanceof ItemBlock ib) {
-			return EmiPort.getBlockRegistry().getNameForObject(ib.field_150939_a);
-		} else {
-			return EmiPort.getItemRegistry().getNameForObject(item);
-		}
-	}
-
-	public static @Nullable String getId(ItemStack stack) {
-		String s = getIdInner(stack);
-		if (s != null && s.contains(":")) {
-			String[] parts = s.split(":");
-			return parts[1];
-		}
-		return null;
-	}
-
-//	public static void setBannerPatterns(ItemStack stack, NBTTagList patterns) {
-//		NBTTagCompound tag = stack.getSubCompound("BlockEntityTag");
-//		if (tag == null) {
-//			tag = new NBTTagCompound();
-//			stack.setTagInfo("BlockEntityTag", tag);
-//		}
-//		tag.setTag("Patterns", patterns);
-//	}
 
 	public static List<Text> getItemToolTip(ItemStack stack, ITooltipFlag.TooltipFlags type) {
 		List<String> rawTip = stack.getTooltip(Minecraft.getMinecraft().thePlayer, type.isAdvanced());
