+++ src/main/java/dev/emi/emi/network/FillRecipeC2SPacket.java
@@ -5,6 +5,7 @@ import java.util.List;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
 
+import com.rewindmc.retroemi.RetroEMICommonUtils;
 import shim.net.minecraft.inventory.ClickType;
 import net.minecraft.network.PacketBuffer;
 import net.minecraft.util.ResourceLocation;
@@ -140,7 +141,7 @@ public class FillRecipeC2SPacket implements EmiPacket {
 					if (gotten != stack.stackSize) {
 						if (gotten > 0) {
 							stack.stackSize = gotten;
-							RetroEMI.offerOrDrop(player, stack);
+							RetroEMICommonUtils.offerOrDrop(player, stack);
 						}
 						return;
 					} else {
@@ -159,7 +160,7 @@ public class FillRecipeC2SPacket implements EmiPacket {
 							}
 							s.putStack(stack);
 						} else {
-							RetroEMI.offerOrDrop(player, stack);
+							RetroEMICommonUtils.offerOrDrop(player, stack);
 						}
 					}
 				}
@@ -172,7 +173,7 @@ public class FillRecipeC2SPacket implements EmiPacket {
 				}
 			} finally {
 				for (ItemStack stack : rubble) {
-					RetroEMI.offerOrDrop(player, stack);
+					RetroEMICommonUtils.offerOrDrop(player, stack);
 				}
 			}
 		}
@@ -225,7 +226,7 @@ public class FillRecipeC2SPacket implements EmiPacket {
 				return grabbed;
 			}
 			ItemStack r = rubble.get(i);
-			if (RetroEMI.canCombine(stack, r)) {
+			if (RetroEMICommonUtils.canCombine(stack, r)) {
 				int wanted = amount - grabbed;
 				if (r.stackSize <= wanted) {
 					grabbed += r.stackSize;
@@ -245,7 +246,7 @@ public class FillRecipeC2SPacket implements EmiPacket {
 				continue;
 			}
 			ItemStack st = s.getStack();
-			if (RetroEMI.canCombine(stack, st)) {
+			if (RetroEMICommonUtils.canCombine(stack, st)) {
 				int wanted = amount - grabbed;
 				if (st.stackSize <= wanted) {
 					grabbed += st.stackSize;
